requirements.txt
Django==5.2.6
Pillow>=8.0.0
psycopg2-binary>=2.8.6
gunicorn>=20.1.0
whitenoise>=5.3.0
python-docx>=0.8.11
openpyxl>=3.0.9
python-pptx>=0.6.21
reportlab>=3.6.0
PyPDF2>=2.10.0